refactor: change interface file name for better consistency
